refactor external-video-player
Thumbnail hitchhiking started to work, but after jumping on a external video, you cannot use the primary player's slider to seek the video. But when the player enters external video from a normal slide, the slider works well. 
Obviously the synchronise function stops to work if the react-player is not re-rendered..need to find where the external video is synchronised with the primary player.